+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Be carefull when updating and check the changelog!
 ### Planned
 - Full dynamic programming support for the `TextbookOptimizationPipeline`, including a Postgres-specific algorithm
 
+### Updates
+- Translating a query into relational algebra now retains ordering information and works for general set queries
+
 ---